Add relation between entities Book and Loan
The entity Loan have a foreign key that creates a relation with Book entity. For this, it was used the @OneToMany annotation. Also, some other validations to saveLoan have been added, such as number of books borrowed without being returned to the library and only can create book loans for books that actually exist and have an are available